move store to shared
