+++ src/settings/index.js
@@ -1,7 +1,7 @@
 import './site.scss';
 import SettingComponent from 'settings/components/setting';
 import settingReducer from 'settings/reducers/setting';
-import { createStore } from 'store';
+import { createStore } from 'shared/store';
 
 const store = createStore(settingReducer);
 let settingComponent = null;
